+++ BACKEND/package.json
@@ -2,12 +2,12 @@
   "dependencies": {
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
-    "dotenv": "^16.4.7",
+    "dotenv": "^16.5.0",
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.10.1",
+    "mongoose": "^8.14.1",
     "multer": "^1.4.5-lts.2",
-    "nodemon": "^3.1.9",
+    "nodemon": "^3.1.10",
     "route": "^0.2.5"
   }
 }
